requirements.txt
annotated-types==0.7.0
anyio==4.9.0
certifi==2025.1.31
click==8.1.8
Deprecated==1.2.18
distro==1.9.0
exceptiongroup==1.2.2
fastapi==0.115.12
filelock==3.18.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
limits==5.0.0
MarkupSafe==3.0.2
mpmath==1.3.0
networkx==3.4.2
numpy==2.2.4
openai==1.75.0
packaging==24.2
pillow==11.2.1
pydantic==2.11.3
pydantic_core==2.33.1
python-dotenv==1.1.0
python-multipart==0.0.20
slowapi==0.1.9
sniffio==1.3.1
starlette==0.46.2
sympy==1.13.1
torch==2.6.0
torchvision==0.21.0
tqdm==4.67.1
typing-inspection==0.4.0
typing_extensions==4.13.2
uvicorn==0.34.1
wrapt==1.17.2
tensorflow==2.15.0
